+++ scripts/npm/prepublish/index.js
@@ -1,13 +1,17 @@
 const path = require( 'path' );
+const rootDir = path.join( __dirname, '..', '..', '..' );
+const envvars = require( path.join( rootDir, 'config/environment' ) ).envvars;
 const fs = require( 'fs' );
-const readdir = require( 'fs-readdir-promise' );
-const Promise = require( 'bluebird' );
-const semver = require( 'semver' );
-const logSymbols = require( 'log-symbols' );
-const util = require( './lib' );
 const isTravis = require( 'is-travis' );
-const componentsDir = path.join( __dirname, '..', '..', '..', 'src' );
+const logSymbols = require( 'log-symbols' );
+const Promise = require( 'bluebird' );
+const readdir = require( 'fs-readdir-promise' );
+const semver = require( 'semver' );
+const util = require( './lib' );
+
+const componentsDir = path.join( rootDir, 'src' );
 let componentsToPublish = [];
+let UNDEFINED;
 
 // Check git's status.
 util.getGitStatus( './' )
@@ -54,13 +58,12 @@ function handleError( msg ) {
 function handleGitStatus( result ) {
   if ( !result.stdout && !result.stderr ) {
     util.printLn.info( 'Git working directory is clean.' );
-  } else {
-    if ( util.option.dryrun ) {
-      return util.printLn.warning(
-        'Your working directory isn\'t clean but this is a dry ' +
+  } else if ( util.option.dryrun ) {
+    util.printLn.warning(
+      'Your working directory isn\'t clean but this is a dry ' +
         'run so I’ll continue...'
-      );
-    }
+    );
+  } else {
     util.printLn.error(
       'Git working directory is not clean. Commit your work before publishing.'
     );
@@ -70,7 +73,7 @@ function handleGitStatus( result ) {
 
 function checkCredentials( result ) {
   // Travis gets its credentials from .travis.yml
-  if ( isTravis ) return;
+  if ( isTravis ) return '';
   if ( util.option.dryrun ) {
     return util.printLn.warning(
       'I\'m not verifying your npm permissions because this is a dry run.'
@@ -84,7 +87,7 @@ function checkoutMaster() {
   // Travis operates in a detached head state so checkout the master branch.
   if ( isTravis ) {
     util.printLn.info(
-      'Checking out ' + process.env.GH_PROD_BRANCH + ' branch...'
+      'Checking out ' + envvars.GH_PROD_BRANCH + ' branch...'
     );
     return util.git.checkoutMaster();
   }
@@ -95,7 +98,7 @@ function checkoutMaster() {
     );
   }
   return util.git.checkBranch().then( function( result ) {
-    if ( !process.env.GH_PROD_BRANCH || !process.env.GH_DEV_BRANCH ) {
+    if ( !envvars.GH_PROD_BRANCH || !envvars.GH_DEV_BRANCH ) {
       util.printLn.error(
         'Publishing to npm from you local machine isn’t recommended but if ' +
         'you’d like to do it anyway, define GH_PROD_BRANCH and ' +
@@ -103,10 +106,10 @@ function checkoutMaster() {
       );
       process.exit( 1 );
     }
-    if ( result.stdout.trim() !== process.env.GH_PROD_BRANCH ) {
+    if ( result.stdout.trim() !== envvars.GH_PROD_BRANCH ) {
       util.printLn.error(
-        'You’re not on the ' + process.env.GH_PROD_BRANCH +
-        ' branch. Merge your changes into ' + process.env.GH_PROD_BRANCH +
+        'You’re not on the ' + envvars.GH_PROD_BRANCH +
+        ' branch. Merge your changes into ' + envvars.GH_PROD_BRANCH +
         ' before publishing.'
       );
       process.exit( 1 );
@@ -129,14 +132,14 @@ function filterComponents( components ) {
 }
 
 function compareVersionNumber( component ) {
-  if ( component.indexOf( 'cf-' ) !== 0 ) return;
+  if ( component.indexOf( 'cf-' ) !== 0 ) return '';
 
-  const manifest = componentsDir + '/' + component + '/package.json',
-      localVersion = JSON.parse(
-        fs.readFileSync( manifest, 'utf8' )
-      ).version;
+  const manifest = componentsDir + '/' + component + '/package.json';
+  // eslint-disable-next-line no-sync
+  const manifestFile = fs.readFileSync( manifest, 'utf8' );
+  const localVersion = JSON.parse( manifestFile ).version;
 
-  return util.getNpmVersion( component ).then( function( data ) {
+  return util.getNpmVersion( component ).then( data => {
     const npmVersion = data['dist-tags'].latest;
     if ( semver.gt( localVersion, npmVersion ) ) {
       util.printLn.success(
@@ -156,17 +159,19 @@ function compareVersionNumber( component ) {
     } else {
       util.printLn.info( component + ' remains ' + npmVersion, true );
     }
-  } ).catch( function( err ) {
-    if ( /returned 404/.test( err ) ) {
+    return {};
+  } ).catch( err => {
+    if ( ( /returned 404/ ).test( err ) ) {
       util.printLn.success( component + ': ' + localVersion, true );
       return {
         name:       component,
         newVersion: localVersion,
-        oldVersion: undefined
+        oldVersion: UNDEFINED
       };
     }
     util.printLn.error( err );
     process.exit( 1 );
+    return {};
   } );
 }
 
@@ -180,25 +185,25 @@ function compareMasterVersionNumber() {
 }
 
 function buildComponents( components ) {
-  let newVersion;
   const diffs = [];
   const masterComponent = components.pop();
 
   /* TODO: Fix bug that results in some entries in the components array to be
      blank. For now, filter them out. */
-  componentsToPublish = components.filter( function( component ) {
+  componentsToPublish = components.filter( component => {
     let diff;
     if ( component ) {
 
       /* While we're iterating, keep track of each component's semver diff.
          If there's no old version it means it's a new component and CF should
          get a minor bump instead of whatever the actual diff is. */
-      diff = !component.oldVersion ? 'minor' : semver.diff(
+      diff = component.oldVersion ? semver.diff(
         component.oldVersion, component.newVersion
-      );
+      ) : 'minor';
       diffs.push( diff );
-      return component.name !== undefined;
+      return component.name !== UNDEFINED;
     }
+    return false;
   } );
 
   // If no components were updated, check if the master component was updated.
@@ -225,9 +230,9 @@ function buildComponents( components ) {
     process.exit( 1 );
   }
 
-  /* Sort the diffs and increment CF by whatever the first (largest)
-     increment is */
-  newVersion = semver.inc( masterComponent.old, diffs.sort().shift() );
+  // Sort the diffs and increment CF by the first (largest) increment.
+  const newVersion = semver.inc( masterComponent.old, diffs.sort().shift() );
+
   util.printLn.success(
     util.pkg.name + ' will be published: ' + masterComponent.old + ' -> ' +
     newVersion + '. See https://goo.gl/cZvnnL.'
@@ -253,7 +258,9 @@ function updateManifest() {
       'I did not update package.json because this is a dry run.'
     );
   }
+  // eslint-disable-next-line no-sync
   fs.writeFileSync( 'package.json', JSON.stringify( util.pkg, null, 2 ) );
+  return '';
 }
 
 function updateChangelog() {
@@ -304,16 +311,16 @@ function push( result ) {
 function publishComponents( result ) {
   if ( result && result.stdout ) util.printLn.console( result.stdout );
   if ( util.option.dryrun ) {
-    return util.printLn.warning(
+    util.printLn.warning(
       'I did not publish anything to npm because this is a dry run.'
     );
+  } else {
+    if ( !componentsToPublish.length ) return UNDEFINED;
+    const components = componentsToPublish.map( component => component.name );
+    util.printLn.info( 'Publishing ' + components.join( ', ' ) + ' to npm...' );
+    return Promise.all( components.map( util.publish ) );
   }
-  if ( !componentsToPublish.length ) return;
-  const components = componentsToPublish.map( function( component ) {
-    return component.name;
-  } );
-  util.printLn.info( 'Publishing ' + components.join( ', ' ) + ' to npm...' );
-  return Promise.all( components.map( util.publish ) );
+  return UNDEFINED;
 }
 
 function finish( result ) {