+++ scripts/npm/prepublish/lib/git.js
@@ -1,8 +1,9 @@
 const childProcess = require( 'child-process-promise' );
+const envvars = require( '../../../../config/environment' ).envvars;
 
 const git = {
   checkoutMaster: function() {
-    return childProcess.exec( 'git checkout ' + process.env.GH_PROD_BRANCH );
+    return childProcess.exec( 'git checkout ' + envvars.GH_PROD_BRANCH );
   },
   checkBranch: function() {
     return childProcess.exec( 'git rev-parse --abbrev-ref HEAD' );
@@ -15,28 +16,28 @@ const git = {
     return childProcess.exec( 'git tag -a ' + version + ' -m "' + version + '"' );
   },
   push: function( remote ) {
-    if ( process.env.GH_TOKEN ) {
+    if ( envvars.GH_TOKEN ) {
       remote = remote.match( /github\.com.+/ )[0];
       return Promise.all( [
         childProcess.exec(
-          'git push "https://' + process.env.GH_TOKEN + '@' + remote + '" ' +
-          process.env.GH_PROD_BRANCH
+          'git push "https://' + envvars.GH_TOKEN + '@' + remote + '" ' +
+          envvars.GH_PROD_BRANCH
         ),
         childProcess.exec(
-          'git push "https://' + process.env.GH_TOKEN + '@' + remote + '" ' +
-          process.env.GH_PROD_BRANCH + ':' + process.env.GH_DEV_BRANCH
+          'git push "https://' + envvars.GH_TOKEN + '@' + remote + '" ' +
+          envvars.GH_PROD_BRANCH + ':' + envvars.GH_DEV_BRANCH
         ),
         childProcess.exec(
-          'git push "https://' + process.env.GH_TOKEN + '@' + remote +
+          'git push "https://' + envvars.GH_TOKEN + '@' + remote +
           '" --tags'
         )
       ] );
     }
     return Promise.all( [
-      childProcess.exec( 'git push ' + remote + ' ' + process.env.GH_PROD_BRANCH ),
+      childProcess.exec( 'git push ' + remote + ' ' + envvars.GH_PROD_BRANCH ),
       childProcess.exec(
-        'git push ' + remote + ' ' + process.env.GH_PROD_BRANCH + ':' +
-        process.env.GH_DEV_BRANCH
+        'git push ' + remote + ' ' + envvars.GH_PROD_BRANCH + ':' +
+        envvars.GH_DEV_BRANCH
       ),
       childProcess.exec( 'git push ' + remote + ' --tags' )
     ] );