+++ scripts/npm/prepublish/lib/print.js
@@ -12,7 +12,7 @@ const options = require( './getArgs' );
  */
 function printMsg( type, msg, indent ) {
   return console.log(
-    indentString( logSymbols[type] + ' ' + msg, ' ', indent ? 4 : 2 )
+    indentString( logSymbols[type] + ' ' + msg, indent ? 4 : 2 )
   );
 }
 
@@ -20,13 +20,16 @@ const printLn = {};
 
 [ 'success', 'warning', 'error', 'info' ].forEach( function( type ) {
   printLn[type] = function( msg, indent ) {
-    options.silent ? function() {} : printMsg( type, msg, indent );
+    if ( !options.silent ) {
+      printMsg( type, msg, indent );
+    }
   };
 } );
 
 printLn.console = function( msg ) {
-  options.silent ? function() {} :
-    console.log( chalk.dim( indentString( msg, ' ', 8 ) ) );
+  if ( !options.silent ) {
+    console.log( chalk.dim( indentString( msg, 8 ) ) );
+  }
 };
 
 module.exports = printLn;