+++ src/cf-atomic-component/src/components/AtomicComponent.js
@@ -10,8 +10,6 @@
 
    ========================================================================== */
 
-'use strict';
-
 const assign = require( '../utilities/object-assign' ).assign;
 const bind = require( '../utilities/function-bind' ).bind;
 const classList = require( '../utilities/dom-class-list' );
@@ -26,7 +24,7 @@ const isFunction = require( '../utilities/type-checkers' ).isFunction;
  * necessary methods to properly instantiatie component.
  *
  * @param {HTMLElement} element - The element to set as the base element.
- * @param {Object} attributes -  Hash of attributes to set on base element.
+ * @param {Object} attributes - Hash of attributes to set on base element.
  */
 function AtomicComponent( element, attributes ) {
   this.element = element;
@@ -52,8 +50,8 @@ assign( AtomicComponent.prototype, Events, classList, {
    * Function used to process class modifiers. These should
    * correspond with BEM modifiers.
    *
-   * @param {Object} attributes -  Hash of attributes to set on base element.
-   * @param {Object} atomicComponent -  Base component.
+   * @param {Object} attributes - Hash of attributes to set on base element.
+   * @param {Object} atomicComponent - Base component.
    */
   processModifiers: function() {
     if ( !this.modifiers ) {
@@ -85,9 +83,9 @@ assign( AtomicComponent.prototype, Events, classList, {
    */
   ensureElement: function() {
     if ( !this.element ) { // eslint-disable-line no-negated-condition, inline-comments, max-len
-      var attrs = assign( {}, this.attributes );
+      const attrs = assign( {}, this.attributes );
       attrs.id = this.id || this.u_id;
-      if ( this.className ) attrs['class'] = this.className;
+      if ( this.className ) attrs.class = this.className;
       this.setElement( document.createElement( this.tagName ) );
       this.setElementAttributes( attrs );
     } else {
@@ -112,15 +110,18 @@ assign( AtomicComponent.prototype, Events, classList, {
     return this;
   },
 
+  // TODO Fix complexity issue
+  /* eslint-disable complexity */
   /**
    * Function used to set the cached DOM elements.
    *
    * @returns {Object} Hash of event names and cached elements.
    */
   setCachedElements: function() {
-    var key;
-    var ui = assign( {}, this.ui );
-    var element;
+    const ui = assign( {}, this.ui );
+    let key;
+    let element;
+
     for ( key in ui ) {
       if ( ui.hasOwnProperty( key ) ) {
         element = this.element.querySelectorAll( ui[key] );
@@ -137,6 +138,7 @@ assign( AtomicComponent.prototype, Events, classList, {
 
     return ui;
   },
+  /* eslint-enable complexity */
 
   /**
    * Function used to remove the base element from the DOM
@@ -159,7 +161,7 @@ assign( AtomicComponent.prototype, Events, classList, {
   /**
    * Function used to set the attributes on an element.
    *
-   * @param {Object} attributes -  Hash of attributes to set on base element.
+   * @param {Object} attributes - Hash of attributes to set on base element.
    */
   setElementAttributes: function( attributes ) {
     let property;
@@ -171,6 +173,8 @@ assign( AtomicComponent.prototype, Events, classList, {
     }
   },
 
+  // TODO Fix complexity issue
+  /* eslint-disable complexity */
   /**
    * Function used to up event delegation on the base element.
    * Using Dom-delegate library to enable this functionality.
@@ -179,31 +183,35 @@ assign( AtomicComponent.prototype, Events, classList, {
    * @returns {AtomicComponent} An instance.
    */
   delegateEvents: function( events ) {
-    var key;
-    var method;
-    var match;
-    var delegateEventSplitter = /^(\S+)\s*(.*)$/;
+    const delegateEventSplitter = /^(\S+)\s*(.*)$/;
+    let key;
+    let method;
+    let match;
 
     events = events || ( events = this.events );
     if ( !events ) return this;
     this.undelegateEvents();
     this._delegate = new Delegate( this.element );
     for ( key in events ) {
-      method = events[key];
-      if ( isFunction( this[method] ) ) method = this[method];
-      if ( !method ) continue;
-      match = key.match( delegateEventSplitter );
-      this.delegate( match[1], match[2], bind( method, this ) );
+      if ( {}.hasOwnProperty.call( events, key ) ) {
+        method = events[key];
+        if ( isFunction( this[method] ) ) method = this[method];
+        if ( method ) {
+          match = key.match( delegateEventSplitter );
+          this.delegate( match[1], match[2], bind( method, this ) );
+        }
+      }
     }
     this.trigger( 'component:bound' );
 
     return this;
   },
+  /* eslint-enable complexity */
 
   /**
    * Function used to set the attributes on an element.
    *
-   * @param {string} eventName -  Event in which to listen for.
+   * @param {string} eventName - Event in which to listen for.
    * @param {string} selector - CSS selector.
    * @param {Function} listener - Callback for event.
    * @returns {AtomicComponent} An instance.
@@ -247,12 +255,12 @@ assign( AtomicComponent.prototype, Events, classList, {
  * Function used to set the attributes on an element.
  * and unbind events.
  *
- * @param {Object} attributes -  Hash of attributes to set on base element.
+ * @param {Object} attributes - Hash of attributes to set on base element.
  * @returns {Function} Extended child constructor function.
  */
 AtomicComponent.extend = function( attributes ) {
 
- /**
+/**
  * Function used as constructor in order to establish inheritance
  * chain.
  * @returns {AtomicComponent} An instance.
@@ -284,11 +292,11 @@ AtomicComponent.extend = function( attributes ) {
  * @returns {Array} List of AtomicComponent instances.
  */
 AtomicComponent.init = function() {
-  var elements = document.querySelectorAll( this.selector );
-  var element;
-  var components = [];
+  const elements = document.querySelectorAll( this.selector );
+  const components = [];
+  let element;
 
-  for ( var i = 0; i < elements.length; ++i ) {
+  for ( let i = 0; i < elements.length; ++i ) {
     element = elements[i];
     if ( element.hasAttribute( 'data-bound' ) === false ) {
       components.push( new this( element ) );