+++ src/cf-atomic-component/src/utilities/object-assign/index.js
@@ -8,8 +8,6 @@
 
    ========================================================================== */
 
-'use strict';
-
 /**
  * @param {object} object - JavaScript object.
  * @returns {boolean} True if object is a plain JavaScript object.
@@ -18,13 +16,15 @@ function _isPlainObject( object ) {
   return Object.prototype.toString.call( object ) === '[object Object]';
 }
 
+// TODO Fix complexity issue
+/* eslint-disable complexity */
 /**
-* Copies properties of all sources to the destination object overriding its own
-* existing properties. When assigning from multiple sources, fields of every
-* next source will override same named fields of previous sources.
-*
-* @param {Object} destination object.
-* @returns {Object} assigned destination object.
+ * Copies properties of all sources to the destination object overriding its own
+ * existing properties. When assigning from multiple sources, fields of every
+ * next source will override same named fields of previous sources.
+ *
+ * @param {Object} destination object.
+ * @returns {Object} assigned destination object.
 */
 function assign( destination ) {
   destination = destination || {};
@@ -44,6 +44,7 @@ function assign( destination ) {
 
   return destination;
 }
+/* eslint-enable complexity */
 
 // Expose public methods.
 module.exports = { assign: assign };