+++ src/cf-expandables/src/cf-expandables.js
@@ -1,6 +1,0 @@
-const Expandable = require( './Expandable' );
-
-// polyfill for ie9 compatibility
-require( 'classlist-polyfill' );
-
-Expandable.init();